Se separa adapter de holder
